scripts: point `xxscreeps import` at a payload of the default shard
The command reads a payload JSON, bundling one exported from the vanilla
default shard, instead of scraping @screeps/launcher's world dump. The
launcher and lokijs dependencies retire with the scraper.

A payload carries terrain and objects, so the launcher's four bots, their
code and their spawns do not survive the switch; `import` seeds an unowned
world and `manage bot add` places players into it.
