+++ .changeset/import-payload-command.md
@@ -1,0 +1,5 @@
+---
+"xxscreeps": patch
+---
+
+`xxscreeps import` reads a payload of the bundled default shard; `@screeps/launcher` retires.
