+++ packages/xxscreeps/.npmignore
@@ -4,6 +4,6 @@
 dist/driver/.cache
 dist/shard.json
 dist/tsconfig.tsbuildinfo
-shard.json
+/shard.json
 test
 tsconfig.json
