+++ packages/xxscreeps/package.json
@@ -25,7 +25,6 @@
     "@isolated-vm/experimental": "^0.0.12",
     "@jridgewell/trace-mapping": "^0.3.31",
     "@loaderkit/resolve": "^1.0.6",
-    "@screeps/launcher": "4.1.0",
     "@types/convert-source-map": "^2.0.3",
     "@xxscreeps/lodash3": "workspace:*",
     "@xxscreeps/pathfinder": "0.4.4",
@@ -47,7 +46,6 @@
     "koa-conditional-get": "3.0.0",
     "koa-passport": "4.1.4",
     "koa-router": "10.1.1",
-    "lokijs": "1.5.12",
     "node-fetch": "3.3.1",
     "openid": "2.0.14",
     "passport-steam": "1.0.18",
@@ -73,7 +71,6 @@
     "@types/koa-conditional-get": "~2.0.0",
     "@types/koa-passport": "~4.0.2",
     "@types/koa-router": "~7.4.4",
-    "@types/lokijs": "~1.5.8",
     "@types/node": "~24.12.2",
     "@types/node-fetch": "~2.6.4",
     "@types/openid": "~2.0.2",
