+++ packages/xxscreeps/scripts/scripts-test.ts
@@ -1,16 +1,24 @@
+import type { Payload } from './payload.js';
 import type { Shard } from 'xxscreeps/engine/db/index.js';
+import type { RoomObject } from 'xxscreeps/game/object.js';
 import type { Room } from 'xxscreeps/game/room/index.js';
 import type { Terrain } from 'xxscreeps/game/terrain.js';
 import * as assert from 'node:assert';
+import * as fs from 'node:fs/promises';
 import { Fn } from 'xxscreeps/functional/fn.js';
 import { instanceOfPredicate } from 'xxscreeps/functional/predicate.js';
 import * as C from 'xxscreeps/game/constants/index.js';
 import { makeSignedRoomName, parseSignedRoomName } from 'xxscreeps/game/room/name.js';
 import { flushUsers } from 'xxscreeps/game/room/room.js';
 import { StructureController } from 'xxscreeps/mods/classic/controller/controller.js';
+import { StructureExtractor } from 'xxscreeps/mods/classic/mineral/extractor.js';
+import { Mineral } from 'xxscreeps/mods/classic/mineral/mineral.js';
+import { StructureKeeperLair } from 'xxscreeps/mods/classic/source/keeper-lair.js';
+import { Source } from 'xxscreeps/mods/classic/source/source.js';
 import { deterministicRandomForTesting } from 'xxscreeps/test/fixtures.js';
 import { instantiateTestShard } from 'xxscreeps/test/import.js';
 import { describe, test } from 'xxscreeps/test/index.js';
+import { importPayload } from './payload.js';
 import { generateRoom, generateSector } from './room-gen.js';
 
 interface SideInfo {
@@ -308,5 +316,28 @@ describe('scripts/room-gen', () => {
 		// The punched ring room sits on the edge of both sectors and carries both centers.
 		assert.deepStrictEqual([ ...terrain.get('W30N25')?.sectors ?? [] ].sort(), [ 'W25N25', 'W35N25' ]);
 		assert.ok(terrain.get('W25N25')?.sectorControl, 'the first sector keeps its control record');
+	});
+});
+
+describe('scripts/payload', () => {
+	// The world `xxscreeps import` ships is only as good as the codecs registered when it was
+	// exported. A codec that drops or renames its marker fails here instead of on a user's install.
+	test('decodes the bundled default shard', async () => {
+		const file = new URL('../../scripts/data/shard.json', import.meta.url);
+		const payload = JSON.parse(await fs.readFile(file, 'utf8')) as Payload;
+		const { rooms } = importPayload(payload);
+		const countOf = function(predicate: (object: RoomObject) => boolean) {
+			return Fn.pipe(
+				Fn.transform(rooms, room => room['#objects']),
+				$$ => Fn.filter($$, predicate),
+				$$ => Fn.accumulate($$, () => 1),
+			);
+		};
+		assert.strictEqual(rooms.length, 121);
+		assert.strictEqual(countOf(instanceOfPredicate(StructureController)), 72);
+		assert.strictEqual(countOf(instanceOfPredicate(Source)), 141);
+		assert.strictEqual(countOf(instanceOfPredicate(Mineral)), 81);
+		assert.strictEqual(countOf(instanceOfPredicate(StructureKeeperLair)), 32);
+		assert.strictEqual(countOf(instanceOfPredicate(StructureExtractor)), 9);
 	});
 });
