+++ packages/xxscreeps/xxscreeps.js
@@ -13,7 +13,7 @@ const commands = {
 	cli: './dist/cli/cli.js',
 	eval: './dist/cli/eval.js',
 	export: './dist/scripts/export.js',
-	import: './dist/scripts/scrape-world.js',
+	import: './dist/scripts/import.js',
 	main: './dist/engine/service/main.js',
 	manage: './dist/scripts/manage.js',
 	processor: './dist/engine/service/processor.js',
